% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_get_mun_history.R
\name{add_past}
\alias{add_past}
\title{Add the past to original data}
\usage{
add_past(x, mutations)
}
\arguments{
\item{x}{The original data as tibble.}

\item{mutations}{Mutation file}
}
\description{
The past is searched by the mHistId.x of the origial file
}
\examples{
mutations <- swcGetMutations()

t <-
  dplyr::filter(mutations, mId.y == 293)

t_1 <- dplyr::filter(t, mHistId.y == max(mHistId.y))

t_past <- add_past(t, mutations)

}
