% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{data}
\name{cantons}
\alias{cantons}
\title{Canton data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 4 columns.
}
\usage{
cantons
}
\description{
Canton data
}
\keyword{datasets}
