% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swcGetMapping.R
\name{computeMunList}
\alias{computeMunList}
\title{The .y values of the argument contain the newly added, the .x values
the dropped column identifiers.  This function verifies that after each
.y value there is a corresponding .x value, and vice versa, and that the
mutation IDs are in nondecreasing order.  The return value is the list of
municipality IDs after all transformations have been applied.}
\usage{
computeMunList(mun.mut.m)
}
\description{
Work around performance issues in plyr:
}
\details{
Split input:

Assign to each x even sequence values, to each y odd sequence values:

Mingle, order by municipality (=value) and sequence number:

Municipality groups: Compute "group change points" and end of group:

Mutation IDs must be nondecreasing, and sequence numbers must be increasing
by one in each group (i.e., everywhere except perhaps at group boundaries)

Return value: All municipalities that have not expired at the end of their
group.
}

