% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{data}
\name{district_mutations}
\alias{district_mutations}
\title{District mutation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 308 rows and 13 columns.
}
\usage{
district_mutations
}
\description{
District mutation data
}
\keyword{datasets}
