% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_historicized_commune_data}
\alias{get_historicized_commune_data}
\title{Import historic commune data from the web}
\usage{
get_historicized_commune_data(force = FALSE)
}
\arguments{
  \item{force}{If \code{TRUE}, reload data from the web.}
}
\value{
A named list with the components \code{canton}, \code{district} and
  \code{municipality}, each component contains a data frame.
}
\description{
This functon returns historic commune.  If necessary,
\code{\link{read_historicized_commune_data}} is called.  The data is cached
with the help of the \link{R.cache} package.
}
\seealso{
\code{\link{read_historicized_commune_data}}
}

