% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swcGetMapping.R
\name{meltMutations}
\alias{meltMutations}
\title{The list of mutations is processed in the order of the mutation ID, which
is composed of mutation date and mutation number. (Conversation with Ernst
Oberholzer end of March 2013.)  All records with the same mutation ID
form a mutation.}
\usage{
meltMutations(mun.mut, hist)
}
\description{
The list of mutations is processed in the order of the mutation ID, which
is composed of mutation date and mutation number. (Conversation with Ernst
Oberholzer end of March 2013.)  All records with the same mutation ID
form a mutation.
}

