% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{data}
\name{municipality_mutations}
\alias{municipality_mutations}
\title{Municipality mutation data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5798 rows and 15 columns.
}
\usage{
municipality_mutations
}
\description{
Municipality mutation data
}
\keyword{datasets}
