% Generated by roxygen2 (4.0.0): do not edit by hand
\name{read_historicized_commune_data}
\alias{read_historicized_commune_data}
\title{Read historic commune data from the web}
\usage{
read_historicized_commune_data()
}
\description{
This functon reads historic commune data from the web and creates a data
frame for each file (canton, district, municipality) in it See also:
\url{http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}

