% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_get_mun_history.R
\name{swc_get_mun_history}
\alias{swc_get_mun_history}
\title{Get municipality mutation history}
\usage{
swc_get_mun_history(munId)
}
\arguments{
\item{munId}{Municipality Id as integer}
}
\description{
The history goes back as far as the mutation number becocmes 1000
  (first time registration)
}
\examples{
waedenswil_history <- swc_get_mun_history(293)

}
