% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swc_get_mun_merges.R
\name{swc_get_mun_merges}
\alias{swc_get_mun_merges}
\title{Get municipality merges for a certain year and cantona}
\usage{
swc_get_mun_merges(year = NULL, canton = NULL)
}
\arguments{
\item{year}{The year where the merges took place}

\item{canton}{The abbreviation as characger
of a canton from which the merges should be returned.
If NULL, all merges are returned}
}
\description{
Get municipality merges for a certain year and cantona
}
\examples{
zuerich_merges <- swc_get_mun_merges(year = 2019, canton = "ZH")

}
