% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipality_state.R
\name{swc_get_municipality_state}
\alias{swc_get_municipality_state}
\title{Existing municipalities in a specified year}
\usage{
swc_get_municipality_state(year, canton = NULL)
}
\arguments{
\item{year}{Year of interest (integer).}

\item{canton}{Canton abbreviation as character (e.g. "GE", "ZH", "TI", etc.) to focus on.
If left `NULL` (default) all cantons are considered.}
}
\value{
Tibble containing municipality numbers (BFS-numbers) and names of existing
municipalities in the given year.
}
\description{
This function produces a tibble containing the names and BFS-numbers
(official municipality numbers) of all existing municipalities in a given year.
Filtering by canton is supported.
}
\examples{
swc_get_municipality_state(1987)
swc_get_municipality_state(2000, "ZH")
}
