% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swcReadData}
\alias{swcReadData}
\title{Import historic commune data from the web}
\usage{
swcReadData()
}
\value{
A named list with the components \code{canton}, \code{district} and
  \code{municipality}, each component contains a data frame.
}
\description{
This functon reads historic commune data from the web and creates a data
frame for each file (canton, district, municipality) in it.
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}

