% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{SwissBirths}
\alias{SwissBirths}
\title{Total births per municipality in Switzerland between 1969 and 2012}
\source{
STAT-TAB, section "Lebendgeburten nach institutionellen Gliederungen, Geschlecht und Staatsangehörigkeit des Kindes und nach Altersklasse der Mutter".

  STAT-TAB URL: \url{http://www.pxweb.bfs.admin.ch/Database/German_01\%20-\%20Bev\%C3\%B6lkerung/01.2\%20-\%20Bev\%C3\%B6lkerungsstand\%20und\%20-bewegung/01.2\%20-\%20Bev\%C3\%B6lkerungsstand\%20und\%20-bewegung.asp?lang=1&prod=01&secprod=2&openChild=true}.

  File link: \url{http://www.pxweb.bfs.admin.ch/Database/German_01\%20-\%20Bev\%C3\%B6lkerung/01.2\%20-\%20Bev\%C3\%B6lkerungsstand\%20und\%20-bewegung/px-d-01-2D02.px}.
}
\description{
Each row contains the number of births for a municipality and
  a year.  Not all municipalities are present for each year due to
  changes in the definition of the municipalities.
}
\author{
Swiss Federal Statistical Office
}

