% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{package}
\name{SwissCommunes-package}
\alias{SwissCommunes-package}
\alias{SwissCommunes}
\title{Functions for working with the historicized list of municipalities of
Switzerland}
\description{
The mutations of Swiss municipalities are tracked in a machine-readable
mutation list.  This allows guessing the reference time for any given list of
municipality IDs, and to construct mappings between two lists of municipality
IDs.  The package provides functions to import and work with this mutation
list.
}
\details{
\tabular{ll}{
  Package: \tab SwissCommunes\cr
  Type: \tab Package\cr
  Version: \tab 0.0-5\cr
  Date: \tab 2014-09-22\cr
  License: \tab GPL-3\cr
}
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}
\author{
Kirill Müller

Maintainer: Kirill Müller <kirill.mueller@ivt.baug.ethz.ch>

URL: \url{https://github.com/krlmlr/SwissCommunes}

Issue tracker: \url{https://github.com/krlmlr/SwissCommunes/issues}
}
\keyword{Switzerland}
\keyword{communes}
\keyword{historicized}
\keyword{municipalities}
