% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{SwissHCD}
\alias{SwissHCD}
\title{Historic commune data, imported from the web}
\value{
A named list with the components \code{canton}, \code{district} and
  \code{municipality}, each component contains a data frame.
}
\description{
This data set is an import of the historic commune data to a data
frame for each file (canton, district, municipality) in it.
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}

