% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/zzx-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{swcGetMapping}
\title{Deprecated functions}
\usage{
swcGetMapping(swc = NULL, ids.from, ids.to)
}
\description{
These functions are deprecated in favor of better alternatives.
}
