% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swcCheckData}
\alias{swcCheckData}
\title{Check internal assumptions on historic commune data}
\usage{
swcCheckData(swc = swcGetData())
}
\arguments{
  \item{swc}{Data as read by the function
  \code{swcReadData}, defaults to \code{swcGetData()}.}
}
\value{
Invisible named list with the following elements:

\describe{
  \item{\code{mutationsWithNonUniqueAdmissionNumbers}}{All municipality
    mutations where the admission number occurs at least in one other
    mutation}
  \item{\code{mutationSequencesWithDecreasingDate}}{Mutation sequences
    where the date decreases}
}
}
\description{
This functon checks several assumptions made for the historic commune data,
  mostly for internal purposes:
}
\details{
\itemize{
  \item Each admission number has less than 5 records (with the exception
    of the initial admission that creates all communes at the earliest
    reported state)
  \item Admission numbers are roughly increasing by date, except for
    differences of one day
  \item The \code{mHist} column is a surrogate key
}
}

