% Generated by roxygen2 (4.0.2): do not edit by hand
\name{swcGetData}
\alias{swcGetData}
\title{Import historic commune data from the web}
\usage{
swcGetData(force = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, reload data from the web.}
}
\value{
A named list with the components \code{canton}, \code{district} and
  \code{municipality}, each component contains a data frame.
}
\description{
This functon returns historic commune data.  If necessary,
\code{\link{swcReadData}} is called.  The data is cached
with the help of the \link{R.cache} package.
}
\seealso{
\code{\link{swcReadData}}
}

