% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{package}
\name{SwissCommunes-package}
\alias{SwissCommunes-package}
\alias{SwissCommunes}
\title{Functions for working with the historicized list of municipalities of
Switzerland}
\description{
The mutations of Swiss municipalities are tracked in a machine-readable
mutation list.  This allows guessing the reference time for any given list of
municipality IDs, and to construct mappings between two lists of municipality
IDs.  The package provides functions to import and work with this mutation
list.

Contains historicized municipality data for
    Switzerland from 1960 onwards, from the "Historisiertes
    Gemeindeverzeichnis" of the Swiss Federal Statistical Office.
}
\details{
\tabular{ll}{
  Package: \tab SwissCommunes\cr
  Type: \tab Package\cr
  Version: \tab 0.0-5\cr
  Date: \tab 2014-09-22\cr
  License: \tab GPL-3\cr
}
}
\examples{
cantons
district_mutations
municipality_mutations
}
\references{
\url{http://www.bfs.admin.ch/bfs/portal/de/index/infothek/nomenklaturen/blank/blank/gem_liste/02.html}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cynkra.github.io/SwissCommunes}
  \item \url{https://github.com/cynkra/SwissCommunes}
  \item Report bugs at \url{https://github.com/cynkra/SwissCommunes/issues}
}

}
\author{
Kirill Müller

Maintainer: Kirill Müller <kirill.mueller@ivt.baug.ethz.ch>

URL: \url{https://github.com/krlmlr/SwissCommunes}

Issue tracker: \url{https://github.com/krlmlr/SwissCommunes/issues}

\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Tobias Schieferdecker
  \item Thomas Knecht
}

Other contributors:
\itemize{
  \item Christoph Sax [contributor]
  \item Swiss Federal Statistical Office (SFSO), Swiss Statistics Web site [copyright holder]
}

}
\keyword{Switzerland}
\keyword{communes}
\keyword{historicized}
\keyword{municipalities}
