% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissCommunes-package.R
\docType{data}
\name{SwissPop}
\alias{SwissPop}
\title{Census population by household size per commune in Switzerland between 1970 and 2000}
\source{
STAT-TAB, section 40.1 "Eidgenössische Volkszählung/1970--2000",
  item "Privathaushalte nach Haushaltsgrösse und Region".

  STAT-TAB URL: \url{http://www.pxweb.bfs.admin.ch/Database/German_40\%20-\%20Eidgen\%C3\%B6ssische\%20Volksz\%C3\%A4hlung/40.1\%20-\%201970-2000/40.1\%20-\%201970-2000.asp?lang=1&prod=40&secprod=1&openChild=true}.

  File link: \url{http://www.pxweb.bfs.admin.ch/Database/German_40\%20-\%20Eidgen\%C3\%B6ssische\%20Volksz\%C3\%A4hlung/40.1\%20-\%201970-2000/px-d-40-1A01.px}.
}
\usage{
SwissPop
}
\description{
Each row contains the number of households for a municipality,
  a year and a household size.
}
\author{
Swiss Federal Statistical Office
}

